# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Topic {
  id: number;
  slug: string;
  title: string;
  weekNumber: number;
  /** @nullable */
  blurb?: string | null;
}

export interface Lecture {
  id: number;
  topicId: number;
  title: string;
  weekNumber: number;
  /** Short Markdown lecture text (the baseline / minimum-detail version). The frontend lets users select passages and send them to the tutor. */
  body: string;
  /**
     * Medium-length version with more explanation and more examples. Null if not yet generated.
     * @nullable
     */
  bodyMedium?: string | null;
  /**
     * Long version with the most explanation and the most examples. Null if not yet generated.
     * @nullable
     */
  bodyLong?: string | null;
}

export interface LectureRef {
  id: number;
  title: string;
  topicId: number;
}

export type AssignmentSummaryKind = typeof AssignmentSummaryKind[keyof typeof AssignmentSummaryKind];


export const AssignmentSummaryKind = {
  homework: 'homework',
  test: 'test',
  midterm: 'midterm',
  final: 'final',
} as const;

export type AssignmentSummaryStatus = typeof AssignmentSummaryStatus[keyof typeof AssignmentSummaryStatus];


export const AssignmentSummaryStatus = {
  not_started: 'not_started',
  in_progress: 'in_progress',
  submitted: 'submitted',
} as const;

export interface AssignmentSummary {
  id: number;
  kind: AssignmentSummaryKind;
  title: string;
  weekNumber: number;
  problemCount: number;
  isTimed: boolean;
  /** @nullable */
  timeLimitMinutes?: number | null;
  status: AssignmentSummaryStatus;
  /** @nullable */
  bestScore?: number | null;
  /** @nullable */
  lastAttemptId?: number | null;
}

export interface Week {
  weekNumber: number;
  title: string;
  /** @nullable */
  summary?: string | null;
  lectures: LectureRef[];
  assignments: AssignmentSummary[];
}

export type AssignmentKind = typeof AssignmentKind[keyof typeof AssignmentKind];


export const AssignmentKind = {
  homework: 'homework',
  test: 'test',
  midterm: 'midterm',
  final: 'final',
} as const;

export interface Problem {
  id: number;
  position: number;
  prompt: string;
  topicId: number;
  /** @nullable */
  topicTitle?: string | null;
  /** @nullable */
  hint?: string | null;
}

export interface Assignment {
  id: number;
  kind: AssignmentKind;
  title: string;
  weekNumber: number;
  isTimed: boolean;
  /** @nullable */
  timeLimitMinutes?: number | null;
  /** @nullable */
  instructions?: string | null;
  problems: Problem[];
}

export type AttemptStateStatus = typeof AttemptStateStatus[keyof typeof AttemptStateStatus];


export const AttemptStateStatus = {
  in_progress: 'in_progress',
  submitted: 'submitted',
} as const;

export interface SavedAnswer {
  problemId: number;
  answer: string;
  /** @nullable */
  keystrokeCount?: number | null;
  /** @nullable */
  eraseCount?: number | null;
}

export interface AttemptState {
  id: number;
  assignmentId: number;
  status: AttemptStateStatus;
  startedAt: string;
  /** @nullable */
  submittedAt?: string | null;
  /** @nullable */
  deadlineAt?: string | null;
  answers: SavedAnswer[];
}

export interface KeystrokeTrace {
  keystrokeCount: number;
  eraseCount: number;
  bulkInsertCount?: number;
  longestBulkInsertChars?: number;
  rewriteSegments?: number;
  durationMs: number;
}

export interface AnswerInput {
  problemId: number;
  answer: string;
  trace: KeystrokeTrace;
}

export interface AnswerSaved {
  ok: boolean;
}

export interface ProblemResult {
  problemId: number;
  correct: boolean;
  userAnswer?: string;
  correctAnswer?: string;
  explanation: string;
}

export interface DetectionResult {
  /** @nullable */
  problemId: number | null;
  /** 0..1 probability the text is AI-generated */
  aiScore: number;
  aiFlagged: boolean;
  /** 0..1 probability the user is rewording AI output (based on keystroke pattern) */
  diachronicScore: number;
  diachronicFlagged: boolean;
  rationale: string;
}

export interface AttemptResult {
  attemptId: number;
  score: number;
  total: number;
  percent: number;
  perProblem: ProblemResult[];
  detection: DetectionResult[];
}

export interface PracticeSessionInput {
  /** @nullable */
  weekNumber?: number | null;
  /** @nullable */
  topicId?: number | null;
  /**
     * When set, practice is scoped to this graded assignment's topics (and disjoint from its graded problems)
     * @nullable
     */
  assignmentId?: number | null;
  tutorEnabled: boolean;
  focusOnWeaknesses?: boolean;
  /**
     * Starting difficulty 1-5; if omitted defaults to 2.0
     * @nullable
     */
  initialDifficulty?: number | null;
}

export interface PracticeSession {
  id: number;
  tutorEnabled: boolean;
  difficulty: number;
  /** @nullable */
  weekNumber: number | null;
  /** @nullable */
  topicId?: number | null;
  /** @nullable */
  assignmentId?: number | null;
  focusOnWeaknesses?: boolean;
}

export interface NextProblemInput {
  /**
     * Free-form user request, e.g. "give me harder probability problems"
     * @nullable
     */
  request?: string | null;
  /** @nullable */
  topicId?: number | null;
}

export interface PracticeProblem {
  id: number;
  prompt: string;
  topicId: number;
  topicTitle?: string;
  difficulty: number;
}

export interface PracticeAnswerInput {
  problemId: number;
  answer: string;
  trace: KeystrokeTrace;
}

export interface PracticeGrade {
  problemId: number;
  correct: boolean;
  correctAnswer?: string;
  explanation: string;
  /** Extensive, in-depth coaching feedback on the student's reasoning (richer than explanation) */
  feedback: string;
  /**
     * Surgically precise, analytics-based pointer on what to focus on before the graded assignment
     * @nullable
     */
  focusPointer?: string | null;
  newDifficulty: number;
  /** @nullable */
  tutorTip?: string | null;
}

export interface TutorAskInput {
  /** @nullable */
  sessionId?: number | null;
  message: string;
  /** @nullable */
  selectedLectureText?: string | null;
  /** @nullable */
  lectureId?: number | null;
  /** @nullable */
  problemId?: number | null;
}

export interface TutorReply {
  text: string;
  /** @nullable */
  audioUrl?: string | null;
}

export interface DetectionScanInput {
  text: string;
  trace: KeystrokeTrace;
}

export interface AnalyticsSummary {
  officialAverage: number;
  practiceAccuracy: number;
  attemptsCount: number;
  practiceCount: number;
  streakDays: number;
  /** @nullable */
  strongestTopic?: string | null;
  /** @nullable */
  weakestTopic?: string | null;
}

export type TopicAnalyticsStrengthLabel = typeof TopicAnalyticsStrengthLabel[keyof typeof TopicAnalyticsStrengthLabel];


export const TopicAnalyticsStrengthLabel = {
  strong: 'strong',
  solid: 'solid',
  developing: 'developing',
  weak: 'weak',
  untested: 'untested',
} as const;

export interface TopicAnalytics {
  topicId: number;
  topicTitle: string;
  weekNumber: number;
  attempts: number;
  accuracy: number;
  strengthLabel: TopicAnalyticsStrengthLabel;
}

export type ActivityItemKind = typeof ActivityItemKind[keyof typeof ActivityItemKind];


export const ActivityItemKind = {
  practice: 'practice',
  assignment: 'assignment',
} as const;

export interface ActivityItem {
  id: number;
  kind: ActivityItemKind;
  title: string;
  at: string;
  /** @nullable */
  score?: number | null;
  /** @nullable */
  topicTitle?: string | null;
}

export interface AnalyticsReport {
  generatedAt: string;
  narrative: string;
  strengths: string[];
  weaknesses: string[];
  recommendations: string[];
}

export type CourseOverviewTotals = {
  assignmentsCompleted: number;
  assignmentsTotal: number;
  practiceCount: number;
};

export interface CourseOverview {
  title: string;
  weeks: Week[];
  totals: CourseOverviewTotals;
}