# lib/api-zod/src/generated/types/practiceGrade.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeGrade {
  problemId: number;
  correct: boolean;
  correctAnswer?: string;
  explanation: string;
  /** Extensive, in-depth coaching feedback on the student's reasoning (richer than explanation) */
  feedback: string;
  /**
     * Surgically precise, analytics-based pointer on what to focus on before the graded assignment
     * @nullable
     */
  focusPointer?: string | null;
  newDifficulty: number;
  /** @nullable */
  tutorTip?: string | null;
}