# lib/api-zod/src/generated/types/practiceSessionInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PracticeSessionInput {
  /** @nullable */
  weekNumber?: number | null;
  /** @nullable */
  topicId?: number | null;
  /**
     * When set, practice is scoped to this graded assignment's topics (and disjoint from its graded problems)
     * @nullable
     */
  assignmentId?: number | null;
  tutorEnabled: boolean;
  focusOnWeaknesses?: boolean;
  /**
     * Starting difficulty 1-5; if omitted defaults to 2.0
     * @nullable
     */
  initialDifficulty?: number | null;
}